feat: :sparkles: added simple kit
