+++ apps/frontend/.env.local.example
@@ -1,1 +1,13 @@
-NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID=
+
+WALLETCONNECT_PROJECT_ID=
+SEPOLIA_ZERO_DEV_PROJECT_ID=
+SEPOLIA_BUNDLER_RPC=https://rpc.zerodev.app/api/v2/bundler/${SEPOLIA_ZERO_DEV_PROJECT_ID}
+SEPOLIA_PAYMASTER_RPC=https://rpc.zerodev.app/api/v2/paymaster/${SEPOLIA_ZERO_DEV_PROJECT_ID}
+
+
+# USED IN APP
+NEXT_PUBLIC_ZERO_DEV_PROJECT_ID=D=${SEPOLIA_ZERO_DEV_PROJECT_ID}
+NEXT_PUBLIC_BUNDLER_RPC=${SEPOLIA_PAYMASTER_RPC}
+NEXT_PUBLIC_PAYMASTER_RPC=${SEPOLIA_PAYMASTER_RPC}
+NEXT_PUBLIC_ZERO_DEV_PROJECT_ID=${SEPOLIA_ZERO_DEV_PROJECT_ID}
+NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID=${WALLETCONNECT_PROJECT_ID}
