+++ apps/frontend/package.json
@@ -21,12 +21,15 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@repo/ui": "workspace:*",
     "@tanstack/react-query": "^5.59.13",
+    "@zerodev/magic-account": "^0.0.30",
+    "@zerodev/social-validator": "^5.1.3",
     "date-fns": "^2.30.0",
     "jotai": "^2.10.0",
     "lucide-react": "^0.395.0",
     "next": "14.2.8",
     "react": "^18",
     "react-dom": "^18",
+    "react-error-boundary": "^4.0.13",
     "ts-node": "^10.9.2",
     "viem": "2.x",
     "wagmi": "^2.12.17"
