+++ apps/frontend/src/modules/wallet/index.ts
